Enum pupil::lexer::Token

pub enum Token<'a> {
    Unk(&'a str),
    Lit(Value),
    Op(Operator),
    Var(&'a str),
    Open(&'a str),
    Comma,
    Close,
}

Supported tokens types.

Variants

Unk

Unknown token.

It’s the caller’s responsibility to handle this with an error of some kind.

This will swallow the entire remainder of the input, such that the tokenization finishes on next iteration.

Lit

Value literal token.

Negative literals are Token::Op(Operator::Sub) followed by a positive literal.

Op

Operator token.

Var

Variable token.

Alphanumeric characters only. Not followed by a (.

Open

Function token.

Alphanumeric characters only. Implicitly followed by a (.

Comma

Comma token ,.

Used to provide multiple arguments to a function.

Close

Function closing token ).